Different controller added; fixes #18
This closes controller issue new controllers added for admin user & client.
